+++ src/feature/attendance/api/index.ts
@@ -1,3 +1,3 @@
 export * from "./hooks";
 export * from "./types";
-export type { AttendanceStatus } from "@/feature/session/api";
+export type { AttendanceStatus } from "@/shared/api";
