+++ src/feature/attendance/index.ts
@@ -1,3 +1,3 @@
 export * from "./api";
-export type { AttendanceStatus } from "@/feature/session/api";
+export type { AttendanceStatus } from "@/shared/api";
 export * from "./hooks";
