+++ src/feature/session/api/types.ts
@@ -1,31 +1,4 @@
-/**
- * 출결 상태
- */
-export type AttendanceStatus =
-  | "present"
-  | "absent"
-  | "late"
-  | "left_early"
-  | "annual"
-  | "leave"
-  | "no_session"
-  | "no_attendance";
-
-/**
- * 세션 정보
- */
-export interface Session {
-  id: number;
-  bootcampId: number;
-  classDate: string;
-  unitNo: number;
-  periodStartDate: string;
-  periodEndDate: string;
-  attendance?: {
-    status: AttendanceStatus;
-    userId: number;
-  };
-}
+export type { AttendanceStatus, Session } from "@/shared/api";
 
 /**
  * 세션 등록 요청
